Rewrite the paste command
