+++ cloudnet/src/main/resources/lang/english.properties
@@ -311,6 +311,7 @@ module-cloudflare-delete-dns-record-for-service=DNS entry was deleted on domain 
 module-report-command-report-description=Creates a file with all important data for a report or for develop modules for CloudNet
 module-report-command-report-post-success=The report file is now created. You can find the file at this path: %file%
 module-report-command-paste-description=Pastes the full service console output to the configured haste server
+module-report-command-paste-service-not-found=That service doesn't exist
 module-report-command-paste-success=The data has been uploaded on %url%
 module-report-command-paste-failed=Failed to upload the data on %url%
 module-report-create-record-start=Starting creation of a new record for %service% in directory %file%
