+++ cloudnet/src/main/resources/lang/german.properties
@@ -311,6 +311,7 @@ module-cloudflare-delete-dns-record-for-service=DNS Eintrag auf der Domain %doma
 module-report-command-report-description=Erstellt eine Datei, welche alle wichtigen Informationen zur Node beinhaltet, diese hilft beim Entwickeln und bei Fehlermeldungen!
 module-report-command-report-post-success=Der Report wurde erfolgreich erstellt. Er ist unter %file% zu finden
 module-report-command-paste-description=Postet die verfügbaren Zeilen der Konsoleausgabe des Services auf den in der Konfiguration des Report-Moduls eingestellten Hasteserver
+module-report-command-paste-service-not-found=Dieser Service existiert nicht
 module-report-command-paste-success=Die Daten sind nun verfügbar unter folgender Url: %url%
 module-report-command-paste-failed=Beim Hochladen der Daten auf %url% ist ein Fehler aufgetreten
 module-report-create-record-start=Es wird ein neuer Eintrag für den Service %service% im Verzeichnis %file% erstellt
